comments + refactor

+++ src/generative_models/gm_params.jl
@@ -1,16 +1,14 @@
 abstract type AbstractGMParams end
 
+# The most basic generative model parameters
 @with_kw struct GMParams <: AbstractGMParams
     n_trackers::Int = 4
     distractor_rate::Real = 4.0
     init_pos_spread::Real = 300.0
-    area_height::Int64 = 800
-    area_width::Int64 = 800
+    area_height::Float64 = 800.0
+    area_width::Float64 = 800.0
     dot_radius::Float64 = 20.0
     targets::Vector{Bool} = zeros(8)
-
-    # # rfs parameters
-    # record_size::Int = 100 # number of associations
 end
 
 function load(::Type{GMParams}, path; kwargs...)
@@ -18,6 +16,7 @@ function load(::Type{GMParams}, path; kwargs...)
 end
 
 
+# Polygon/hierarchical version of GMParams
 @with_kw struct HGMParams <: AbstractGMParams
     n_trackers::Int64 = 4
     distractor_rate::Float64 = 4.0
